Simplify metadata rewrite pipeline
